pyproject.toml
[project]
name = "interview-qa-extractor"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dagster>=1.10.0",
    "dagster-cloud>=1.10.0",
    "extract_interview_questions"
]

[dependency-groups]
dev = [
    "dagster-webserver>=1.10.0",
    "mypy>=1.13.0",
    "pytest-dotenv>=0.5.2",
    "pytest>=8.3.3",
    "ruff>=0.7.4",
    "ollama>=0.4.7",
    "pypdf2>=3.0.1",
    "langchain-ollama>=0.2.3",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.17",
    "pypdf>=5.3.0",
    "youtube-transcript-api>=0.6.3",
]

[tool.uv.sources]
extract_interview_questions = { workspace = true }

[tool.uv.workspace]
members = ["workspaces/*"]

[tool.mypy]
strict = true
files = ["workspaces/**/src"]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
env_files = [".env"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
pythonpath = [
  "workspaces/extract_interview_questions",
  "workspaces/extract_interview_questions/src",
]
